Move AnnotationInput to Components/AnnotationInput.tsx
