+++ labextension/src/components/Components.tsx
@@ -15,39 +15,9 @@
  */
 
 import * as React from 'react';
-import { createStyles, makeStyles, useTheme } from '@material-ui/core/styles';
-import { Button, Switch } from '@material-ui/core';
-import DeleteIcon from '@material-ui/icons/Delete';
-import { RokInput } from './RokInput';
+import { useTheme } from '@material-ui/core/styles';
+import { Switch } from '@material-ui/core';
 import { Input } from './Input';
-
-// https://codeburst.io/my-journey-to-make-styling-with-material-ui-right-6a44f7c68113
-const useStyles = makeStyles(() =>
-  createStyles({
-    label: {
-      color: 'var(--jp-input-border-color)',
-      fontSize: 'var(--jp-ui-font-size2)',
-    },
-    input: {
-      color: 'var(--jp-ui-font-color1)',
-    },
-    focused: {},
-    // notchedOutline: {
-    //     borderWidth: '1px',
-    //     borderColor: 'var(--jp-input-border-color)',
-    // },
-    textField: {
-      width: '100%',
-    },
-    menu: {
-      backgroundColor: 'var(--jp-layout-color1)',
-      color: 'var(--jp-ui-font-color1)',
-    },
-    helperLabel: {
-      color: 'var(--jp-info-color0)',
-    },
-  }),
-);
 
 interface ICollapsablePanel {
   title: string;
@@ -96,94 +66,6 @@ export const CollapsablePanel: React.FunctionComponent<ICollapsablePanel> = prop
           />
         </div>
       </div>
-    </div>
-  );
-};
-
-interface IAnnotationInput {
-  updateValue: Function;
-  deleteValue: Function;
-  annotation: { key: string; value: string };
-  volumeIdx: number;
-  annotationIdx: number;
-  label: string;
-  cannotBeDeleted?: boolean;
-  rokAvailable?: boolean;
-}
-
-export const AnnotationInput: React.FunctionComponent<IAnnotationInput> = props => {
-  const [annotation, setAnnotation] = React.useState({ key: '', value: '' });
-  const classes = useStyles({});
-
-  React.useEffect(() => {
-    // need this to set the annotation when the notebook is loaded
-    // and the metadata is updated
-    setAnnotation({ ...props.annotation });
-  }, [props.annotation]); // Only re-run the effect if props.annotation changes
-
-  const updateKey = (key: string) => {
-    props.updateValue(
-      { ...props.annotation, key: key },
-      props.volumeIdx,
-      props.annotationIdx,
-    );
-  };
-
-  const updateValue = (value: string) => {
-    props.updateValue(
-      { ...props.annotation, value: value },
-      props.volumeIdx,
-      props.annotationIdx,
-    );
-  };
-
-  const valueField =
-    props.rokAvailable && props.annotation.key === 'rok/origin' ? (
-      <RokInput
-        updateValue={updateValue}
-        value={props.annotation.value}
-        label={'Rok URL'}
-        inputIndex={props.volumeIdx}
-        annotationIdx={props.annotationIdx}
-      />
-    ) : (
-      <Input
-        updateValue={updateValue}
-        value={props.annotation.value}
-        label={'Value'}
-        inputIndex={props.volumeIdx}
-        variant="standard"
-      />
-    );
-
-  return (
-    <div className="toolbar">
-      <div style={{ marginRight: '10px', width: '50%' }}>
-        <Input
-          updateValue={updateKey}
-          value={props.annotation.key}
-          label={'Key'}
-          inputIndex={props.volumeIdx}
-          readOnly={props.cannotBeDeleted || false}
-          variant="standard"
-        />
-      </div>
-      <div style={{ width: '50%' }}>{valueField}</div>
-      {!props.cannotBeDeleted ? (
-        <div className="delete-button">
-          <Button
-            variant="contained"
-            size="small"
-            title="Remove Annotation"
-            onClick={_ =>
-              props.deleteValue(props.volumeIdx, props.annotationIdx)
-            }
-            style={{ transform: 'scale(0.9)' }}
-          >
-            <DeleteIcon />
-          </Button>
-        </div>
-      ) : null}
     </div>
   );
 };
